Moved files around, typed stuff
