+++ src/modules/auth/auth.types.ts
@@ -1,0 +1,6 @@
+enum UserTypes {
+  User = 'User',
+  Shelter = 'Shelter',
+}
+
+export type TokenType = string | boolean | UserTypes;
