+++ src/modules/auth/strategies/jwt.strategy.ts
@@ -17,7 +17,7 @@ export class JwtStrategy extends PassportStrategy(Strategy, 'jwt') {
     super({
       jwtFromRequest: ExtractJwt.fromExtractors([
         (request: Request) => {
-          return request?.cookies?.access_token;
+          return request?.cookies?.access_token as string;
         },
         ExtractJwt.fromAuthHeaderAsBearerToken(),
       ]),
